moved menu_component helper and template into view_component directory
