+++ app/helpers/learning_partners_helper.rb
@@ -14,7 +14,7 @@ module LearningPartnersHelper
   private
 
   def edit_partner_item(learning_partner)
-    MenuComponentHelper::MenuItem.new(
+    ViewComponent::MenuComponentHelper::MenuItem.new(
       label: t('button.edit'),
       url: edit_learning_partner_path(learning_partner),
       type: :link
@@ -22,7 +22,7 @@ module LearningPartnersHelper
   end
 
   def activate_partner_item(learning_partner)
-    MenuComponentHelper::MenuItem.new(
+    ViewComponent::MenuComponentHelper::MenuItem.new(
       label: t('learning_partner.activate'),
       url: activate_learning_partner_path(learning_partner),
       type: :link,
@@ -37,7 +37,7 @@ module LearningPartnersHelper
   end
 
   def payment_plan_item(learning_partner)
-    MenuComponentHelper::MenuItem.new(
+    ViewComponent::MenuComponentHelper::MenuItem.new(
       label: t('learning_partner.payment_plan.modify'),
       url: edit_learning_partner_payment_plan_path(learning_partner),
       type: :link,
@@ -46,7 +46,7 @@ module LearningPartnersHelper
   end
 
   def impersonate_partner_item(learning_partner)
-    MenuComponentHelper::MenuItem.new(
+    ViewComponent::MenuComponentHelper::MenuItem.new(
       label: t('impersonation.impersonate'),
       url: impersonation_path(id: learning_partner.id),
       type: :button,
@@ -55,7 +55,7 @@ module LearningPartnersHelper
   end
 
   def deactivate_partner_item(learning_partner)
-    MenuComponentHelper::MenuItem.new(
+    ViewComponent::MenuComponentHelper::MenuItem.new(
       label: t('learning_partner.block'),
       url: deactivate_learning_partner_path(learning_partner),
       type: :link,
