+++ app/helpers/teams_helper.rb
@@ -90,25 +90,25 @@ module TeamsHelper
 
   def team_menu_items(team)
     [
-      MenuComponentHelper::MenuItem.new(
+      ViewComponent::MenuComponentHelper::MenuItem.new(
         label: 'Team dashboard',
         url: dashboard_path(team_id: team),
         type: :link,
         visible: policy(:dashboard).index?
       ),
-      MenuComponentHelper::MenuItem.new(
+      ViewComponent::MenuComponentHelper::MenuItem.new(
         label: 'Edit',
         url: edit_team_path(team),
         type: :link,
         options: { data: { turbo_frame: 'modal' } },
         visible: policy(team).edit?
       ),
-      MenuComponentHelper::MenuItem.new(
+      ViewComponent::MenuComponentHelper::MenuItem.new(
         label: 'View all users',
         url: all_users_team_path(team),
         type: :link
       ),
-      MenuComponentHelper::MenuItem.new(
+      ViewComponent::MenuComponentHelper::MenuItem.new(
         label: 'Generate Report',
         url: new_report_path(team_id: team.id),
         type: :link,
