+++ app/views/view_components/menu_component/_menu_component.html.erb
